+汇智接口提交

+++ core/src/main/java/com/bingo/hotel/spa/intl/core/api/huitravel/bean/price/availability/NightlyRate.java
@@ -2,10 +2,12 @@ package com.bingo.hotel.spa.intl.core.api.huitravel.bean.price.availability;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class NightlyRate {
     private String date;
-    private Integer cost;
+    private BigDecimal cost;
     private boolean status;
     private Integer roomnum;
 }
